requirements.txt
langgraph
langchain
langchain-core
langchain_google_genai
pydantic
python-dotenv